preprocessing PUCP_DGI and update to generalize preprocessing\segmentedPerSRT
A new file extractingSRTeaf.py was created to generate SRT files from EAF files, which are basically XML files. The script to segment videos based on SRT files was updated to generalize to any dataset. Before it had hard coded width and height to work with AEC videos but now those should be passed by parameters so it is supported to have a whole frame of sign language.